Add console script
This commit has a few tweaks rolled into one. First, arg parsing was
moved to __main__.py, leaving start.py as a dummy wrapper. Second,
a console_script was added to make launching angr-management easier
from anywhere.
