+++ angrmanagement/__main__.py
@@ -3,6 +3,11 @@ import os
 import ctypes
 import threading
 import time
+import warnings
+
+def shut_up(*args, **kwargs):
+    return
+warnings.simplefilter = shut_up
 
 
 def check_dependencies():
@@ -34,8 +39,7 @@ def set_app_user_model_id():
             ctypes.windll.shell32.SetCurrentProcessExplicitAppUserModelID(myappid)
 
 
-def main(filepath=None):
-
+def start_management(filepath=None):
     if not check_dependencies():
         sys.exit(1)
 
@@ -83,6 +87,28 @@ def main(filepath=None):
         main_window.load_file(file_to_open)
 
     app.exec_()
+
+def main():
+    import argparse
+    import runpy
+
+    parser = argparse.ArgumentParser(description="angr management")
+    parser.add_argument("-s", "--script", type=str, help="run a python script in the (commandline) angr environment")
+    parser.add_argument("-i", "--interactive", action='store_true', help="interactive (ipython) mode")
+    parser.add_argument("-n", "--no-gui", action='store_true', help="run in headless mode")
+    parser.add_argument("binary", nargs="?", help="the binary to open (for the GUI)")
+
+    args = parser.parse_args()
+
+    if args.script:
+        script_globals = runpy.run_path(args.script)
+    if args.interactive:
+        if args.script:
+            print("Your script's globals() dict is available in the `script_globals` variable.")
+        import IPython
+        IPython.embed(banner1="")
+    if not args.no_gui:
+        start_management(args.binary)
 
 if __name__ == '__main__':
     main()
